New entity reference representation side-by-side
I thought a commit having the old and the new in the same file would be
useful if we need to come back and investigate the differences between
the two entity reference sets in the future.  I've verified that the two
are exactly the same with the exception of the ten listed in the comment
at the bottom.